move wrap_resources_for_execution, use in build_resources
