+++ python_modules/dagster/dagster/core/definitions/graph.py
@@ -537,10 +537,8 @@ class GraphDefinition(NodeDefinition):
         Returns:
             InProcessGraphResult
         """
-        from dagster.core.execution.execute_in_process import (
-            core_execute_in_process,
-            wrap_resources_for_execution,
-        )
+        from dagster.core.execution.build_resources import wrap_resources_for_execution
+        from dagster.core.execution.execute_in_process import core_execute_in_process
         from dagster.core.instance import DagsterInstance
         from .job import JobDefinition
         from .executor import execute_in_process_executor
